feat: refactor GroupRowSelection context and hooks for improved grid handling and selection management
