+++ React/src/GroupRowSelection/context/helpers.ts
@@ -1,0 +1,4 @@
+export const serializeKey = (key: any) =>
+  typeof key === "string" || typeof key === "number"
+    ? String(key)
+    : JSON.stringify(key, Object.keys(key).sort());
